requirements.txt
flask
tensorflow-cpu==2.13.0
numpy
Pillow
werkzeug
gunicorn
